ANFIS code
Adaptive Neuro-Fuzzy Interference System(ANFIS) is a type of artificial neural network. The code for training of this network is uploaded along with the code used for data generation. The data generation is done by using forward kinematics. This generated data trains the ANFIS. Further, the error plot code is also uploaded to generate the plots containing the error in ANFIS network. This is further compared with errors produced in PID control.